fix: change parameters of Communications __init__
First, I corrected the order of arguments when initializing an instance
of the `Communications` class. In `main.py`, arguments are ordered like
the following:
```
+    comm = comms.Communications(USB, SCREEN_DIMENSIONS, FPS, CAM_IDS)
```

However, initially, the arguments for initialization were:
```
-    def __init__(self, usb, camera_id1, camera_id2, SCREEN_DIMENSIONS, FPS):
```

I removed `camera_id1` and `camera_id2` and replaced it with `cam_ids`.
This is arguably better, for several reasons. One, it is not guaranteed
that there will be only two cameras. Two, it is more eloquent, with
multiple cameras, to iterate over them in an array.

`cam_ids` was also moved to be the last argument to mirror the
initialization in `main.py`.

`CAM_IDS` in `main.py` likely needs to be updated to reflect the actual
amount of cameras intended to be used.

Signed-off-by: Lucas Sta Maria <lucas.stamaria@gmail.com>