fix: add buffer and write_to_buffer function for reading serial
Removed a lot of the additional testing/experimenting code.

Stores a character array in a `buffer`, read byte by byte from the
serial. The buffer is passed by reference to the `write_to_buffer`
function.  The `buf_ind` variable keeps track of where the buffer is
currently being written to. It is passed by reference to the
`write_to_buffer` function so that it can be modified after the buffer
has been written to.

If `Serial.available()` causes the while loop to break, the `buffer` and
`buf_ind` variables are discarded.

Signed-off-by: Lucas Sta Maria <lucas.stamaria@gmail.com>