fix: check for valid packets in buffer
`has_packet` checks if there is a terminating byte, denoted by '\255'
(or all bits flipped to 1). If there is, then there exists an end of a
packet. However, if packet verification has to be performed in case a
byte was dropped.

Two potential sources of error arise from a dropped byte. First, the
dropped byte is information in the packet. The resulting packet would
then be too short, and an error would be thrown. Second, the dropped
byte is a terminating byte, meaning the packet would be too long and
overflow the packet char array.

`has_valid_packet` resolves this by checking the position of the
terminating byte. If the terminating byte is not at the expected
location, then it can be assumed that either one of the failing cases
has occured. The buffer should then be reset.

Signed-off-by: Lucas Sta Maria <lucas.stamaria@gmail.com>