fix: make `has_valid_packet` return correct result for packets
The `if` statement previously checked that the terminating character
`'\255'` was at the fourth index (index 3) rather than the third index
(index 2). A fix was made to remove the additional `+ 1` which made it
so.

Additionally, if that `if` statement wasn't fulfilled, the entire
function would return `false` at the first index (index 0). Moving the
`return false` to outside the `for` loop resolves that.

Signed-off-by: Lucas Sta Maria <lucas.stamaria@gmail.com>