fix: move packet int array declaration to outer loop
With every pass of the while loop, the packet data would be erased.
Moving it outside the while loop resolves that. When the serial becomes
unavailable, this also resets the `buffer` and `packet` data. This is
useful, since a disconnect with the Serial might mean bytes being lost.
In order to not use corrupt data, another iteration resets the `buffer`
and `packet` data.

Signed-off-by: Lucas Sta Maria <lucas.stamaria@gmail.com>