doc(mcu.ino): add comments for documentation
Header explains the purpose of the file and provides the GPLv3 license
boilerplate.

Program's constants and definitions are explained with comments.
User-defined functions are explained with comments, so that those
unfamiliar with the program have a better grasp of what is happening.

Made `115200` into the constant `BAUD_RATE` to remove the use of a
potentially unclear magic number.

Signed-off-by: Lucas Sta Maria <lucas.stamaria@gmail.com>